requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.1
django-environ==0.4.5
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
flake8==3.8.3
importlib-metadata==2.0.0
kombu==5.0.2
Markdown==3.2.2
mccabe==0.6.1
prompt-toolkit==3.0.7
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.0
